# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CineConnect API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface UserProfile {
  id: string;
  clerkId: string;
  name: string;
  email: string;
  /** @nullable */
  avatarUrl?: string | null;
  /** Actor | Director | Cameraman | Editor | Music | Writer | Producer | Other */
  role: string;
  /** @nullable */
  bio?: string | null;
  skills: string[];
  /** @nullable */
  location?: string | null;
  /** @nullable */
  portfolioUrl?: string | null;
  portfolioImages: string[];
  isVerified: boolean;
  /** online | offline | away */
  onlineStatus: string;
  /** @nullable */
  averageRating?: number | null;
  totalReviews: number;
  totalProjectsPosted: number;
  createdAt: string;
}

export interface UpdateProfileBody {
  name?: string;
  role?: string;
  bio?: string;
  skills?: string[];
  location?: string;
  portfolioUrl?: string;
  portfolioImages?: string[];
  avatarUrl?: string;
  onlineStatus?: string;
}

export interface UserListResponse {
  users: UserProfile[];
  total: number;
  page: number;
  limit: number;
}

export type FeedStatsRoleBreakdownItem = {
  role: string;
  count: number;
};

export interface FeedStats {
  totalUsers: number;
  roleBreakdown: FeedStatsRoleBreakdownItem[];
  recentJoiners: UserProfile[];
  activeProjects: number;
}

export interface Project {
  id: number;
  title: string;
  description: string;
  requiredRoles: string[];
  /** @nullable */
  location?: string | null;
  isRemote: boolean;
  /** open | closed | in-progress */
  status: string;
  posterUserId: string;
  posterName: string;
  /** @nullable */
  posterAvatarUrl?: string | null;
  posterRole: string;
  applicationCount: number;
  isBookmarked: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface ProjectListResponse {
  projects: Project[];
  total: number;
  page: number;
  limit: number;
}

export interface CreateProjectBody {
  title: string;
  description: string;
  requiredRoles: string[];
  location?: string;
  isRemote?: boolean;
}

export interface ProjectApplication {
  id: number;
  projectId: number;
  applicantUserId: string;
  message: string;
  status: string;
  createdAt: string;
}

export interface ApplyProjectBody {
  message: string;
}

export interface Conversation {
  id: number;
  otherUserId: string;
  otherUserName: string;
  /** @nullable */
  otherUserAvatarUrl?: string | null;
  otherUserRole: string;
  /** @nullable */
  lastMessage?: string | null;
  /** @nullable */
  lastMessageAt?: string | null;
  unreadCount: number;
}

export interface Message {
  id: number;
  conversationId: number;
  senderUserId: string;
  senderName: string;
  /** @nullable */
  senderAvatarUrl?: string | null;
  content: string;
  isRead: boolean;
  createdAt: string;
}

export interface SendMessageBody {
  recipientUserId: string;
  content: string;
}

export interface Review {
  id: number;
  reviewerUserId: string;
  reviewerName: string;
  /** @nullable */
  reviewerAvatarUrl?: string | null;
  revieweeUserId: string;
  /** 1 to 5 */
  rating: number;
  /** @nullable */
  comment?: string | null;
  createdAt: string;
}

export interface ReviewListResponse {
  reviews: Review[];
  /** @nullable */
  averageRating?: number | null;
  total: number;
}

export interface CreateReviewBody {
  rating: number;
  comment?: string;
}

export interface Notification {
  id: number;
  userId: string;
  /** new_message | new_project | profile_view | new_application | new_review */
  type: string;
  title: string;
  body: string;
  isRead: boolean;
  /** @nullable */
  relatedUserId?: string | null;
  /** @nullable */
  relatedProjectId?: number | null;
  createdAt: string;
}

export type ListUsersParams = {
  role?: string;
  location?: string;
  skill?: string;
  search?: string;
  page?: number;
  limit?: number;
};

export type RecordProfileView200 = {
  success: boolean;
};

export type BlockUser200 = {
  success: boolean;
};

export type ListProjectsParams = {
  role?: string;
  location?: string;
  search?: string;
  page?: number;
  limit?: number;
};

export type BookmarkProject200 = {
  bookmarked: boolean;
};

export type GetMessagesParams = {
  before?: number;
  limit?: number;
};

export type GetUnreadCount200 = {
  count: number;
};

export type MarkAllNotificationsRead200 = {
  success: boolean;
};