# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CineConnect API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /messages/conversations/{conversationId}:
    get:
      operationId: getMessages
      tags: [messages]
      summary: Get messages in a conversation
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: integer
        - name: before
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
  /messages/send:
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message (creates conversation if needed)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageBody"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
...
  /notifications/{notificationId}/read:
    post:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a notification as read
      parameters:
        - name: notificationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Marked read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
components:
  schemas:
...
    Message:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        senderUserId:
          type: string
        senderName:
          type: string
        senderAvatarUrl:
          type: ["string", "null"]
        content:
          type: string
        isRead:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - conversationId
        - senderUserId
        - senderName
        - content
        - isRead
        - createdAt
    SendMessageBody:
      type: object
      properties:
        recipientUserId:
          type: string
        content:
          type: string
      required:
        - recipientUserId
        - content
...
    Notification:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        type:
          type: string
          description: "new_message | new_project | profile_view | new_application | new_review"
        title:
          type: string
        body:
          type: string
        isRead:
          type: boolean
        relatedUserId:
          type: ["string", "null"]
        relatedProjectId:
          type: ["integer", "null"]
        createdAt:
          type: string
      required:
        - id
        - userId
        - type
        - title
        - body
        - isRead
        - createdAt